Move protos into one folder, don't rely on .pb.go
