+++ BUILD.bazel
@@ -23,11 +23,9 @@ filegroup(
         "//config:all-srcs",
         "//hack:all-srcs",
         "//images:all-srcs",
-        "//issue_state:all-srcs",
         "//metadata:all-srcs",
+        "//pb:all-srcs",
         "//resultstore:all-srcs",
-        "//state:all-srcs",
-        "//summary:all-srcs",
         "//util/gcs:all-srcs",
     ],
     tags = ["automanaged"],
