+++ cmd/entomologist/BUILD.bazel
@@ -8,7 +8,8 @@ go_library(
     visibility = ["//visibility:private"],
     deps = [
         "//config:go_default_library",
-        "//issue_state:go_default_library",
+        "//pb/config:go_default_library",
+        "//pb/issue_state:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@io_k8s_apimachinery//pkg/util/wait:go_default_library",
@@ -30,8 +31,8 @@ go_test(
     srcs = ["main_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//config:go_default_library",
-        "//issue_state:go_default_library",
+        "//pb/config:go_default_library",
+        "//pb/issue_state:go_default_library",
         "@io_k8s_test_infra//pkg/io:go_default_library",
         "@io_k8s_test_infra//prow/flagutil:go_default_library",
         "@io_k8s_test_infra//prow/github:go_default_library",
