+++ cmd/summarizer/BUILD.bazel
@@ -29,9 +29,9 @@ go_library(
     importpath = "github.com/GoogleCloudPlatform/testgrid/cmd/summarizer",
     visibility = ["//visibility:private"],
     deps = [
-        "//cmd/summarizer/response:go_default_library",
         "//cmd/summarizer/summary:go_default_library",
-        "//summary:go_default_library",
+        "//pb/response:go_default_library",
+        "//pb/summary:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@org_golang_google_grpc//:go_default_library",
     ],
@@ -42,9 +42,9 @@ go_test(
     srcs = ["summarizer_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//cmd/summarizer/response:go_default_library",
-        "//config:go_default_library",
-        "//summary:go_default_library",
+        "//pb/config:go_default_library",
+        "//pb/response:go_default_library",
+        "//pb/summary:go_default_library",
         "@com_github_google_go_cmp//cmp:go_default_library",
     ],
 )
@@ -61,7 +61,6 @@ filegroup(
     srcs = [
         ":package-srcs",
         "//cmd/summarizer/example:all-srcs",
-        "//cmd/summarizer/response:all-srcs",
         "//cmd/summarizer/summary:all-srcs",
     ],
     tags = ["automanaged"],
