+++ cmd/summarizer/example/BUILD.bazel
@@ -9,10 +9,10 @@ go_library(
     importpath = "github.com/GoogleCloudPlatform/testgrid/cmd/summarizer/example",
     visibility = ["//visibility:private"],
     deps = [
-        "//cmd/summarizer/response:go_default_library",
         "//cmd/summarizer/summary:go_default_library",
         "//config:go_default_library",
-        "//summary:go_default_library",
+        "@com_github_googlecloudplatform_testgrid//cmd/summarizer/response:go_default_library",
+        "@com_github_googlecloudplatform_testgrid//summary:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@org_golang_google_grpc//:go_default_library",
         "@org_golang_google_grpc//credentials:go_default_library",
