--- cmd/summarizer/response/BUILD.bazel
@@ -1,44 +1,0 @@
-load("@io_bazel_rules_go//proto:def.bzl", "go_proto_library")
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    embed = [":response_go_proto"],
-    importpath = "github.com/GoogleCloudPlatform/testgrid/cmd/summarizer/response",
-    visibility = ["//visibility:public"],
-)
-
-filegroup(
-    name = "package-srcs",
-    srcs = glob(["**"]),
-    tags = ["automanaged"],
-    visibility = ["//visibility:private"],
-)
-
-filegroup(
-    name = "all-srcs",
-    srcs = [":package-srcs"],
-    tags = ["automanaged"],
-    visibility = ["//visibility:public"],
-)
-
-proto_library(
-    name = "response_proto",
-    srcs = ["types.proto"],
-    visibility = ["//visibility:public"],
-    deps = [
-        "//config:config_proto",
-        "//summary:summary_proto",
-    ],
-)
-
-go_proto_library(
-    name = "response_go_proto",
-    importpath = "github.com/GoogleCloudPlatform/testgrid/cmd/summarizer/response",
-    proto = ":response_proto",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//config:go_default_library",
-        "//summary:go_default_library",
-    ],
-)
