+++ cmd/updater/BUILD.bazel
@@ -23,7 +23,8 @@ go_library(
     deps = [
         "//config:go_default_library",
         "//metadata/junit:go_default_library",
-        "//state:go_default_library",
+        "//pb/config:go_default_library",
+        "//pb/state:go_default_library",
         "//util/gcs:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
         "@com_google_cloud_go//storage:go_default_library",
@@ -51,7 +52,7 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//metadata/junit:go_default_library",
-        "//state:go_default_library",
+        "//pb/state:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
     ],
 )
