--- issue_state/BUILD.bazel
@@ -1,36 +1,0 @@
-load("@io_bazel_rules_go//proto:def.bzl", "go_proto_library")
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-filegroup(
-    name = "package-srcs",
-    srcs = glob(["**"]),
-    tags = ["automanaged"],
-    visibility = ["//visibility:private"],
-)
-
-filegroup(
-    name = "all-srcs",
-    srcs = [":package-srcs"],
-    tags = ["automanaged"],
-    visibility = ["//visibility:public"],
-)
-
-go_library(
-    name = "go_default_library",
-    embed = [":issue_state_go_proto"],
-    importpath = "github.com/GoogleCloudPlatform/testgrid/issue_state",
-    visibility = ["//visibility:public"],
-)
-
-proto_library(
-    name = "issue_state_proto",
-    srcs = ["issue_state.proto"],
-    visibility = ["//visibility:public"],
-)
-
-go_proto_library(
-    name = "issue_state_go_proto",
-    importpath = "github.com/GoogleCloudPlatform/testgrid/issue_state",
-    proto = ":issue_state_proto",
-    visibility = ["//visibility:public"],
-)
