+++ pb/response/BUILD.bazel
@@ -1,0 +1,44 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("@io_bazel_rules_go//proto:def.bzl", "go_proto_library")
+
+proto_library(
+    name = "response_proto",
+    srcs = ["types.proto"],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pb/config:config_proto",
+        "//pb/summary:summary_proto",
+    ],
+)
+
+go_proto_library(
+    name = "response_go_proto",
+    importpath = "github.com/GoogleCloudPlatform/testgrid/pb/response",
+    proto = ":response_proto",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pb/config:go_default_library",
+        "//pb/summary:go_default_library",
+    ],
+)
+
+go_library(
+    name = "go_default_library",
+    embed = [":response_go_proto"],
+    importpath = "github.com/GoogleCloudPlatform/testgrid/pb/response",
+    visibility = ["//visibility:public"],
+)
+
+filegroup(
+    name = "package-srcs",
+    srcs = glob(["**"]),
+    tags = ["automanaged"],
+    visibility = ["//visibility:private"],
+)
+
+filegroup(
+    name = "all-srcs",
+    srcs = [":package-srcs"],
+    tags = ["automanaged"],
+    visibility = ["//visibility:public"],
+)
