--- summary/BUILD.bazel
@@ -1,36 +1,0 @@
-load("@io_bazel_rules_go//proto:def.bzl", "go_proto_library")
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-filegroup(
-    name = "package-srcs",
-    srcs = glob(["**"]),
-    tags = ["automanaged"],
-    visibility = ["//visibility:private"],
-)
-
-filegroup(
-    name = "all-srcs",
-    srcs = [":package-srcs"],
-    tags = ["automanaged"],
-    visibility = ["//visibility:public"],
-)
-
-go_library(
-    name = "go_default_library",
-    embed = [":summary_go_proto"],
-    importpath = "github.com/GoogleCloudPlatform/testgrid/summary",
-    visibility = ["//visibility:public"],
-)
-
-proto_library(
-    name = "summary_proto",
-    srcs = ["summary.proto"],
-    visibility = ["//visibility:public"],
-)
-
-go_proto_library(
-    name = "summary_go_proto",
-    importpath = "github.com/GoogleCloudPlatform/testgrid/summary",
-    proto = ":summary_proto",
-    visibility = ["//visibility:public"],
-)
